fix(scheduler): use lossless f64 conversion for reservation scale factor
apply_new_capacity computed the scale as
    f64::from(new_capacity) / f64::from(total_reserved as u16)
where total_reserved is u32. The intermediate `as u16` would silently
truncate any sum > u16::MAX. The current invariant (`Σ reserved ≤
initial_capacity ≤ u16::MAX`) means we can't hit this today — but a
future code path that increases a reservation post-construction would
silently corrupt the scale factor without any compile-time signal.

f64 represents u32 losslessly (53-bit mantissa); convert directly so
the lossy intermediate cast can never reappear by accident.

+++ model_gateway/src/middleware/scheduler/engine.rs
@@ -399,7 +399,14 @@ impl PriorityScheduler {
             .map(|c| u32::from(self.slot_pool.reserved(*c)))
             .sum();
         if total_reserved > u32::from(new_capacity) && total_reserved > 0 {
-            let scale = f64::from(new_capacity) / f64::from(total_reserved as u16);
+            // total_reserved is u32; converting via `as u16` would
+            // silently truncate any value > u16::MAX. Today the
+            // invariant `Σ reserved ≤ initial_capacity ≤ u16::MAX`
+            // holds (PriorityScheduler::new enforces it and this method
+            // only scales down), but the cast is fragile against future
+            // code paths that might increase a reservation. f64 holds
+            // u32 losslessly (53-bit mantissa > 32 bits).
+            let scale = f64::from(new_capacity) / f64::from(total_reserved);
             for class in Class::ALL {
                 let r = self.slot_pool.reserved(class);
                 let scaled = (f64::from(r) * scale).floor() as u16;
